// src/game.services/inventory.services/inventory.interface.ts
export interface CardMetaData {
    name: string;
    description: string;
    group: string;
    era: string;
    scoreBoost: string;
    healBoost: string;
    skill: string;
    stars: string;
    slot: string;
    tier: string;
    boostCount: string;
    awakenCount: string;
    breakthrough: boolean;
    skillEquipped: boolean;
    artist: string; 
    position: string;
    position2: string;
    rarity: string;
    level: string;
    zodiac: string;
    animal: string;
    experience: string;
    currentExperience: string;
    image: string;
    imageByte?: string;

    uri: string;
    owner: string;
    tokenId: string;
    contractAddress: string;
    id: string;
    listingId: number;
    cardAddress: string;
    uploader: string;
    supply: string;
    quantity: 1
    quantityOwned: string
    lister?: string;
    pricePerToken?: number
    currencyName?: string
    transferred?: string
 };


/**
 * Interface for updating inventory data.
 * @property {uri} the uri of the card saved on IPFS which contains the cards metadata
 * @property {tokenId} the token ID of the card
 * @property {contractAddress} the contract address of the card
 * @property {group} the group of the card
 * @property {slot} the slot of the card
 */
export interface UpdateInventoryData {
  uri: string;
  tokenId: string;
  contractAddress: string;
  group: string;
  slot: string;
  name: string;
}


/**
 * Type representing inventory card data indexed by URI.
 */
export type InventoryCardData = { [uri: string]: CardMetaData };

/**
 * Type representing an array containing two arrays of inventory card data.
 * The first array contains owned and bagged cards, and the second array contains owned and equipped cards.
 */
export type InventoryCards = [InventoryCardData[], InventoryCardData[]];



export interface CardPackData {
  child: boolean;
  currencyName: string;
  description: string;
  endTime: string;
  id: string;
  image: string;
  lister: string;
  listingId: number;
  name: string;
  pricePerToken: number;
  quantity: number;
  startTime: string;
  tokenId: string;
  type: string;
  uploader: string;
  uri: string;
}



export interface UpdateData {
  uri: string;
  tokenId: string;
  contractAddress: string;
  group: string;
  slot: string;
}